requirements.txt
apns-client==0.2.1
appnope==0.1.0
asn1crypto==0.24.0
autopep8==1.4.3
bcrypt==3.1.3
boto3==1.4.5
botocore==1.5.92
certifi==2017.4.17
#cffi==1.10.0
chardet==3.0.3
cookies==2.2.1
cryptography==2.6.1
decorator==4.0.11
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.11
django-background-tasks==1.1.10
django-compat==1.0.14
django-cors-headers==2.0.2
django-debug-toolbar==1.11
django-directmessages==0.9.7
django-extensions==1.7.8
django-extra-fields==0.9
django-filter==1.0.4
django-qr-code==1.1.0
django-silk==3.0.1
django-storages==1.6.5
django-timezone-field==2.0
djangorestframework==3.6.2
djangorestframework-expander==0.2.3
djangorestframework-filters==0.10.1
djangorestframework-gis==0.11.2
docutils==0.14
Faker==0.7.17
gprof2dot==2016.10.13
gunicorn==19.7.1
idna==2.5
#ipdb==0.10.3
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
Jinja2==2.10
jmespath==0.9.3
maas==1.3.2
Markdown==2.6.8
MarkupSafe==1.1.1
oauthlib==2.0.2
olefile==0.44
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
prompt-toolkit==1.0.14
psycopg2==2.7.7
ptyprocess==0.5.1
pycodestyle==2.5.0
pycparser==2.18
pyfcm==1.4.5
Pygments==2.2.0
PyJWT==1.5.0
pyOpenSSL==19.0.0
python-dateutil==2.6.0
python3-openid==3.1.0
pytz==2017.2
PyYAML==3.12
qrcode==5.3
raven==6.3.0
requests==2.17.3
requests-oauthlib==0.8.0
requests-toolbelt==0.9.1
responses==0.5.1
s3transfer==0.1.10
simplegeneric==0.8.1
six==1.10.0
social-auth-core==1.3.0
sqlparse==0.3.0
squareconnect==2.20190327.0
stripe==1.59.0
traitlets==4.3.2
urllib3==1.21.1
wcwidth==0.1.7
whitenoise==3.3.0
zbarlight==1.2